reorganize views, templates
