+++ main/templates/main/dashboard/index.html
@@ -35,7 +35,7 @@
             {% endif %}
           </ul>
         {% else %}
-          <p class="my-3">You're currently not part of any organizations. 
+          <p class="my-3">You're currently not part of any organizations.
             Ask someone to add you to their organization or create a new one below.</p>
           <!-- <a class="btn btn-primary" href="/org/create" role="button">Create New Organization</a> -->
         {% endif %}
