+++ main/templates/main/dashboard/partners/thanks.html
