+++ main/templates/main/organization/home.html
@@ -13,12 +13,7 @@
         <h2 class="text-center">Organization not found!</h2>
         <p class="text-center">Check the URL or search for an organization.</p>
       {% endif%}
-      
-      <!-- <p class=""> Fill this with organization details and helpful links.</p> -->
-        <!--Ideas: - link to form to edit organization details
-          - link to form to add admin users, mod users, and whitelist users
-          - link to moderating page -->
-               
+
       </div>
   </div>
 </div>
